Merge to master

+++ src/metadata.c
@@ -28,7 +28,7 @@ error_type_t write_metadata(uint8_t *data, uint8_t dataSize) {
     return OK;
 }
 
-void override_metadatas(uint8_t offset, void *ptr, size_t size) {
+void override_metadatas(size_t offset, void *ptr, size_t size) {
     nvm_write((void *) &N_storage.metadatas[offset], ptr, size);
 }
