+++ src/ui/ui_passwords.c
@@ -136,8 +136,8 @@ static void confirm_password_deletion(const size_t index) {
         all_passwords = true;
         snprintf(msgBuffer,
                  sizeof(msgBuffer),
-                 "Confirm the deletion\nof all passwords (%d)",
-                 N_storage.metadata_count);
+                 "Confirm the deletion\nof all passwords (%u)",
+                 (uint32_t) N_storage.metadata_count);
     } else {
         // A single password
         all_passwords = false;
